+++ templates/react-native/navigation/DrawerNavigator.tsx
@@ -1,6 +1,7 @@
 import {
   DrawerContentComponentProps,
   DrawerItem,
+  DrawerNavigationProp,
   createDrawerNavigator,
 } from '@react-navigation/drawer';
 import React, { ReactElement } from 'react';
@@ -8,7 +9,15 @@ import { ScrollView, StyleSheet } from 'react-native';
 
 import { useSafeArea } from 'react-native-safe-area-context';
 
-const Drawer = createDrawerNavigator();
+export type DrawerParamList = {
+  default: undefined;
+};
+
+export type DrawerNavigationProps<
+  T extends keyof DrawerParamList = 'default'
+> = DrawerNavigationProp<DrawerParamList, T>;
+
+const Drawer = createDrawerNavigator<DrawerParamList>();
 
 const styles = StyleSheet.create({
   container: {